# tests/unit_tests/fixtures/openapi.yaml
openapi: 3.0.0
info:
  title: V0 Semantic layer API
  version: 1.0.11
  contact:
    name: Андрей Абакин
    url: https://addressbook/search/profile/1763632
    email: abakin.a.v@sberbank.ru
  description: Semantic layer API for EPMPlatform
servers:
  - url: https://ci00163863-epmp-sndbx.apps.dev-terra000004-idm.ocp.delta.sbrf.ru
    description: DEV стенд
  - url: https://ci04763767-semlayer.apps.ift-terra000039-edb.ocp.delta.sbrf.ru
    description: IFT стенд
paths:
...
  /api/v0/tenants/{tenantName}/models/{modelName}/synchronization/composites/:
    post:
      tags:
        - v0
      summary: Create Composites In Database From Meta
      operationId: createCompositesInDatabaseFromMeta
      description: Обновить все view в базе данных
      responses:
        "200":
          description: Обновленные view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/respCopyObject"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/synchronization/composites/{compositeName}:
    post:
      tags:
        - v0
      summary: Create Composite In Database From Meta
      operationId: createCompositeInDatabaseFromMeta
      description: Обновить view в базе данных
      responses:
        "200":
          description: Обновленные view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/synchronizationResult"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
      - $ref: "#/components/parameters/compositeName"
...
  /api/v0/tenants/{tenantName}/models/{modelName}/composites/:
    get:
      parameters:
        - example: 1
          name: page
          description: Страница (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationPageCount"
          in: query
          required: false
        - example: 20
          name: pageSize
          description: Количество строк (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationRowCount"
          in: query
          required: false
        - name: sortDirection
          description: Направление сортировки
          schema:
            $ref: "#/components/schemas/sortDirection"
          in: query
          required: false
      tags:
        - v0
      summary: Get all composites by model name
      operationId: getCompositesByModelName
      responses:
        "200":
          description: Все композиты модели `modelName`
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/compositeGet"
                maxItems: 1000
                uniqueItems: true
        "500":
          $ref: "#/components/responses/unexpectedError"
    post:
      tags:
        - v0
      summary: Create composite by model name
      operationId: createCompositeByModelName
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/compositeGet"
          description: Композит создан в модели `modelName`
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Композит уже существует в модели `modelName`
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/composite"
        description: New composite
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/composites/{compositeName}:
    get:
      tags:
        - v0
      summary: Get composite by model name and composite name
      operationId: getCompositeByModelNameAndCompositeName
      responses:
        "200":
          description: Описание композита с именем `compositeName`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/compositeGet"
        "404":
          description: Не найден композит с именем `compositeName`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
        "500":
          $ref: "#/components/responses/unexpectedError"
    delete:
      tags:
        - v0
      summary: Delete composite by model name and composite name
      operationId: deleteCompositeByModelNameAndCompositeName
      responses:
        "200":
          description: Композит `compositeName` в модели `modelName` удален
        "404":
          description: Композит `compositeName` в модели `modelName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
    patch:
      tags:
        - v0
      summary: Change composite by model name and composite name
      operationId: changeCompositeByModelNameAndCompositeName
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/compositeGet"
          description: Композит `compositeName` в модели `modelName` изменен
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Композит `compositeName` в модели `modelName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/compositeEdit"
    parameters:
      - $ref: "#/components/parameters/compositeName"
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/composites/copyToAnotherModel:
    post:
      tags:
        - v0
      summary: Copy composite to another model
      operationId: copyCompositeToAnotherModel
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/respCopyObject"
          description: Измерение создано в указанных моделях
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
          description: Измерение уже существует в модел `modelName`
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/requestCopyObject"
      description: New model for measure
    parameters:
      - $ref: "#/components/parameters/tenantName"
components:
  responses:
    unexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
  parameters:
    modelName:
      name: modelName
      description: Уникальный идентификатор модели
      schema:
        $ref: "#/components/schemas/modelName"
      in: path
      required: true
...
    compositeName:
      name: compositeName
      description: Уникальный идентификатор композита
      schema:
        $ref: "#/components/schemas/compositeName"
      in: path
      required: true
...
    tenantName:
      name: tenantName
      description: Название тенанта
      schema:
        $ref: "#/components/schemas/tenantName"
      in: path
      required: true
...
  schemas:
    tenantName:
      $ref: "#/components/schemas/objectName32"
...
    compositeName:
      $ref: "#/components/schemas/objectName32"
...
    dataStorageName:
      $ref: "#/components/schemas/objectName64"
    modelName:
      $ref: "#/components/schemas/objectName32"
    objectName32:
      description: Уникальное техническое имя объекта
      type: string
      minLength: 1
      maxLength: 32
      pattern: "^[a-z0-9_]+$"
      example: "objectname32"
...
    schemaName:
      description: Имя схемы базы данных
      type: string
      pattern: "^[a-zA-Z0-9_]+$"
      minLength: 0
      maxLength: 255
      example: "schema_name"
...
    requestCopyObject:
      type: object
      additionalProperties: false
      properties:
        modelNames:
          type: array
          items:
            $ref: "#/components/schemas/modelName"
          maxItems: 1000
          uniqueItems: true
        objects:
          type: array
          items:
            $ref: "#/components/schemas/objectName32"
          maxItems: 1000
          uniqueItems: true
...
    respCopyObject:
      type: array
      maxItems: 1000
      uniqueItems: true
      items:
        type: object
        additionalProperties: false
        properties:
          tenant:
            $ref: "#/components/schemas/tenantName"
          objectName:
            $ref: "#/components/schemas/objectName32"
          modelNames:
            type: array
            items:
              $ref: "#/components/schemas/modelName"
            maxItems: 1000
            uniqueItems: true
          result:
            $ref: "#/components/schemas/result"
          comment:
            pattern: '^[^\n\t]+$'
            type: string
            maxLength: 255
            description: Комментарий
...
    httpValidationError:
      type: object
      additionalProperties: false
      properties:
        detail:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            properties:
              loc:
                type: array
                items:
                  type: string
                  maxLength: 255
                  pattern: '^[^\n\t]+$'
                maxItems: 1000
                uniqueItems: true
              msg:
                $ref: "#/components/schemas/msg"
              type:
                type: string
                maxLength: 255
                pattern: '^[^\n\t]+$'
                description: Тип ошибки
...
    error:
      required:
        - message
      type: object
      additionalProperties: false
      properties:
        message:
          description: A human readable error message
          type: string
          pattern: "^[a-zA-Z0-9_]+$"
          minLength: 1
          maxLength: 255
...
    objectLabels:
      type: array
      maxItems: 100
      uniqueItems: true
      description: |
        Тексты, которые описывают различные сущности системы на разных языках для
          - признаков
          - атрибутов признаков
          - показателей
          - хранилищ данных
          - композитов
          - ...
      items:
        type: object
        additionalProperties: false
        required:
          - language
          - type
        description: Зависимый от языка текст различной длины
        properties:
          language:
            type: string
            maxLength: 5
            description: Язык текста
            example: ru-ru
            pattern: "^[a-z]{2}-[a-z]{2}$"
          type:
            $ref: "#/components/schemas/labelType"
          text:
            type: string
            description: Отображаемый текст
            pattern: '^[^\n\t]+$'
            minLength: 1
            maxLength: 255
...
    composite:
      allOf:
        - $ref: "#/components/schemas/versionInfo"
        - $ref: "#/components/schemas/compositeEdit"
        - type: object
          additionalProperties: false
          required:
            - datasources
            - linkType
            - name
            - fields
          properties:
            name:
              $ref: "#/components/schemas/compositeName"
            schemaName:
              $ref: "#/components/schemas/schemaName"
            aorType:
              $ref: '#/components/schemas/aorType'
    compositeEdit:
      type: object
      additionalProperties: false
      properties:
        labels:
          $ref: "#/components/schemas/objectLabels"
        datasources:
          type: array
          description: Источники данных в композите
          items:
            $ref: "#/components/schemas/datasource"
          maxItems: 1000
          minItems: 1
          uniqueItems: true
        fields:
          type: array
          description: Поля из источников данных в композите
          items:
            $ref: "#/components/schemas/compositeField"
          minItems: 1
          maxItems: 1000
          uniqueItems: true
        linkType:
          type: string
          description: Тип связи источников данных
          enum:
            - LEFT_JOIN
            - INNER_JOIN
            - UNION
            - SELECT
        linkFields:
          type: array
          description: Список полей композита, участвующих в JOIN
          maxItems: 1000
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            properties:
              left:
                $ref: "#/components/schemas/linkField"
              right:
                $ref: "#/components/schemas/linkField"
    compositeGet:
      type: object
      additionalProperties: false
      required:
        - name
        - schemaName
        - fields
        - datasources
        - linkType
      properties:
        aorType:
          $ref: '#/components/schemas/aorType'
        version:
          $ref: "#/components/schemas/version"
        updatedAt:
          $ref: "#/components/schemas/timestamp"
        updatedBy:
          $ref: "#/components/schemas/user"
        name:
          $ref: "#/components/schemas/compositeName"
        schemaName:
          $ref: "#/components/schemas/schemaName"
          deprecated: true
        labels:
          $ref: "#/components/schemas/objectLabels"
        fields:
          type: array
          minItems: 1
          maxItems: 1000
          uniqueItems: true
          items:
            $ref: "#/components/schemas/compositeField"
        dbObjects:
          type: array
          description: Список объектов БД для data storage
          items:
            $ref: "#/components/schemas/dbObject"
          minItems: 0
          maxItems: 1000
        models:
          type: array
          description: Модели, использующие композит, со статусом обработки объекта в нем
          items:
            $ref: "#/components/schemas/modelName"
          maxItems: 1000
          minItems: 0
          uniqueItems: true
        datasources:
          type: array
          description: Источники данных в композите
          items:
            $ref: "#/components/schemas/datasource"
          maxItems: 1000
          minItems: 1
          uniqueItems: true
        linkType:
          type: string
          description: Тип связи источников данных
          enum:
            - LEFT_JOIN
            - INNER_JOIN
            - UNION
            - SELECT
        linkFields:
          type: array
          description: Список полей композита, участвующих в JOIN
          maxItems: 1000
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            properties:
              left:
                $ref: "#/components/schemas/linkField"
              right:
                $ref: "#/components/schemas/linkField"
...
    linkField:
      type: object
      required:
        - datasource
        - datasourceField
      additionalProperties: false
      properties:
        datasource:
          $ref: "#/components/schemas/objectName255"
        datasourceField:
          $ref: "#/components/schemas/objectName255"
...
    sortDirection:
      description: Направление сортировки
      default: asc
      enum:
        - asc
        - desc
      type: string
    paginationPageCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
    paginationRowCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
...
    synchronizationResult:
      type: object
      additionalProperties: false
      properties:
        tenant:
          $ref: "#/components/schemas/tenantName"
        modelName:
          $ref: "#/components/schemas/modelName"
        objectName:
          $ref: "#/components/schemas/dataStorageName"
        comment:
          type: string
          description: Комментарий к синхронизации
          maxLength: 255
          minLength: 0
          pattern: '^[^\n\t]+$'
        status:
          type: string
          description: Статус синхронизации
          enum:
            - SUCCESS
            - IGNORED
            - FAILURE
...
    result:
      description: Результат
      type: string
      enum:
        - SUCCESS
        - PENDING
        - FAILURE
    msg:
      type: string
      maxLength: 8192
      minLength: 1
      description: Сообщение
      example: Комментарий
...
    aorType:
      description: Тип объекта в AOR
      type: string
      enum:
        - sl_model
        - sl_hierarchy
        - sl_dimension
        - sl_measure
        - sl_datastorage
        - sl_databaseobject
        - sl_composite
        - sl_database
        - sl_aigroup
      example: sl_dimension
...
    version:
      description: Версия объекта
      type: integer
      minimum: 1
      maximum: 2147483647
      example: 6
...
    timestamp:
      type: string
      description: Дата и время
      example: "2025-05-06 18:16:15.690"
    user:
      type: string
      description: Пользователь
      pattern: ^[^\n\t]+$
      example: EPMSuperUser
...